+++ CHANGELOG.md
@@ -1,7 +1,13 @@
-## 3.5.2
-* Generate BIOS JSON's file during start-up and after BIOS update
+## 3.5.3
 * Internal:
   * make starting DCSpy client in thread more flexible
+
+## 3.5.2
+* GUI should not be hide when starting when configuration is wrong
+* Generate BIOS JSON's file during start-up and after BIOS update
+* Internal:
+  * move default config into resources directory
+  * check if BIOS JSON's files exist before do generation and compilation
 
 ## 3.5.1
 * Add missing `F-4E-45MC.yaml` - #316 (@emcek)